Add a way for a frontend to generate more complex dwarf location information.  This allows arbitrary code involving DW_OP_plus_uconst and DW_OP_deref.  The scheme allows for easy extention to include, any, or all of the DW_OP_ opcodes.  I thought about just exposing all of them, but, wasn't sure if people wanted the dwarf opcodes exposed in the api.  Is that a layering violation?
With this scheme, the entire existing block scheme used by llvm-gcc
can be switched over to the new scheme.  I think that would be
cleaner, as then the compiler specific bits are not present in llvm
proper.  Before the old code can be yanked however, similar code in
clang would have to be removed.

Next up, more testing.


git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@83120 91177308-0d34-0410-b5e6-96231b3b80d8